Revert "fix bug calling intValue for UnsignedLong offset in NBDVolumeServer"
This reverts commit d9ba36ec3a522aa26731038d20d4fe3134c85f45.

+++ src/main/java/rocky/ctrl/NBDVolumeServer.java
@@ -101,8 +101,7 @@ public class NBDVolumeServer implements Runnable {
             if (RockyStorage.debugPrintoutFlag) {
             	log.info("Reading " + buffer.length + " from " + offset);
             }
-            //storage.read(buffer, offset.intValue()).thenApply($ -> {
-            storage.read(buffer, offset.longValue()).thenApply($ -> {
+            storage.read(buffer, offset.intValue()).thenApply($ -> {
               synchronized (out) {
                 try {
                   out.write(NBD_REPLY_MAGIC_BYTES);
@@ -124,8 +123,7 @@ public class NBDVolumeServer implements Runnable {
             if (RockyStorage.debugPrintoutFlag) {
             	log.info("Writing " + buffer.length + " to " + offset);
             }
-            //storage.write(buffer, offset.intValue()).thenApply($ -> {
-            storage.write(buffer, offset.longValue()).thenApply($ -> {
+            storage.write(buffer, offset.intValue()).thenApply($ -> {
               try {
                 writeReplyHeaderAndFlush(handle);
               } catch (IOException e) {
